Resolução de bugs - findOne e Testes

+++ backend/src/main/java/com/uepb/ControleBiblioteca/repository/JornaisRepository.java
@@ -9,6 +9,6 @@ import com.uepb.ControleBiblioteca.entities.Jornais;
 public interface JornaisRepository extends JpaRepository<Jornais, Integer> {
 
 	Jornais findOne(int id);
-	Jornais update(Jornais jornais, int id);
+	Jornais update(Jornais jornais, Long id);
 	
 }
